refactor(afrl): script name, remove deprecated
