+++ scripts/typescript/afrl-to-new-api/src/scripts/afrl-loader.ts
